+++ contracts/interfaces/callback/IPrimitiveLiquidityCallback.sol
@@ -2,6 +2,15 @@
 pragma solidity 0.8.0;
 
 interface IPrimitiveLiquidityCallback {
-    function allocateCallback(uint deltaX, uint deltaY, bytes calldata data) external;
-    function removeCallback(uint deltaX, uint deltaY, bytes calldata data) external;
+    function allocateCallback(
+        uint256 deltaX,
+        uint256 deltaY,
+        bytes calldata data
+    ) external;
+
+    function removeCallback(
+        uint256 deltaX,
+        uint256 deltaY,
+        bytes calldata data
+    ) external;
 }
