+++ contracts/libraries/Reserve.sol
@@ -12,19 +12,19 @@ library Reserve {
     // An Engine has two reserves of RISKY and RISK-FREE assets, X and Y, and total liquidity shares.
     struct Data {
         // the reserve for the risky asset
-        uint RX1;
+        uint256 RX1;
         // the reserve for the stable asset
-        uint RY2;
+        uint256 RY2;
         // the total supply of liquidity shares
-        uint liquidity;
+        uint256 liquidity;
         // the liquidity available for lending
-        uint float;
+        uint256 float;
         // the liquidity unavailable because it was borrowed
-        uint debt;
+        uint256 debt;
         // oracle items
-        uint cumulativeRisky;
-        uint cumulativeStable;
-        uint cumulativeLiquidity;
+        uint256 cumulativeRisky;
+        uint256 cumulativeStable;
+        uint256 cumulativeLiquidity;
         uint32 blockTimestamp;
     }
 
@@ -38,8 +38,14 @@ library Reserve {
     }
 
     /// @notice Increases one reserve value and decreases the other by different amounts
-    function swap(Data storage reserve, bool addXRemoveY, uint deltaIn, uint deltaOut, uint32 blockTimestamp) internal returns (Data storage) {
-        if(addXRemoveY) {
+    function swap(
+        Data storage reserve,
+        bool addXRemoveY,
+        uint256 deltaIn,
+        uint256 deltaOut,
+        uint32 blockTimestamp
+    ) internal returns (Data storage) {
+        if (addXRemoveY) {
             reserve.RX1 += deltaIn;
             reserve.RY2 -= deltaOut;
         } else {
@@ -50,7 +56,13 @@ library Reserve {
     }
 
     /// @notice Add to both reserves and total supply of liquidity
-    function allocate(Data storage reserve, uint deltaX, uint deltaY, uint deltaL, uint32 blockTimestamp) internal returns (Data storage) {
+    function allocate(
+        Data storage reserve,
+        uint256 deltaX,
+        uint256 deltaY,
+        uint256 deltaL,
+        uint32 blockTimestamp
+    ) internal returns (Data storage) {
         reserve.RX1 += deltaX;
         reserve.RY2 += deltaY;
         reserve.liquidity += deltaL;
@@ -58,7 +70,13 @@ library Reserve {
     }
 
     /// @notice Remove from both reserves and total supply of liquidity
-    function remove(Data storage reserve, uint deltaX, uint deltaY, uint deltaL, uint32 blockTimestamp) internal returns (Data storage) {
+    function remove(
+        Data storage reserve,
+        uint256 deltaX,
+        uint256 deltaY,
+        uint256 deltaL,
+        uint32 blockTimestamp
+    ) internal returns (Data storage) {
         reserve.RX1 -= deltaX;
         reserve.RY2 -= deltaY;
         reserve.liquidity -= deltaL;
@@ -66,26 +84,26 @@ library Reserve {
     }
 
     /// @notice Increases available float to borrow, called when lending
-    function addFloat(Data storage reserve, uint deltaL) internal returns (Data storage) {
+    function addFloat(Data storage reserve, uint256 deltaL) internal returns (Data storage) {
         reserve.float += deltaL;
         return reserve;
     }
 
     /// @notice Reduces available float, taking liquidity off the market, called when claiming
-    function removeFloat(Data storage reserve, uint deltaL) internal returns (Data storage) {
+    function removeFloat(Data storage reserve, uint256 deltaL) internal returns (Data storage) {
         reserve.float -= deltaL;
         return reserve;
     }
 
     /// @notice Reduces float and increases debt of the global reserve, called when borrowing
-    function borrowFloat(Data storage reserve, uint deltaL) internal returns (Data storage) {
+    function borrowFloat(Data storage reserve, uint256 deltaL) internal returns (Data storage) {
         reserve.float -= deltaL;
         reserve.debt += deltaL;
         return reserve;
     }
 
-    /// @notice Increases float and reduces debt of the global reserve, called when repaying a borrow 
-    function repayFloat(Data storage reserve, uint deltaL) internal returns (Data storage) {
+    /// @notice Increases float and reduces debt of the global reserve, called when repaying a borrow
+    function repayFloat(Data storage reserve, uint256 deltaL) internal returns (Data storage) {
         reserve.float += deltaL;
         reserve.debt -= deltaL;
         return reserve;
