+++ contracts/test/engine/EngineAllocate.sol
@@ -12,32 +12,49 @@ contract EngineAllocate {
 
     constructor() {}
 
-    function initialize(address _engine, address _risky, address _stable) public {
-      engine = _engine;
-      risky = _risky;
-      stable = _stable;
+    function initialize(
+        address _engine,
+        address _risky,
+        address _stable
+    ) public {
+        engine = _engine;
+        risky = _risky;
+        stable = _stable;
     }
 
-    function allocateFromMargin(bytes32 pid, address owner, uint dLiquidity, bytes calldata data) public  {
-        IPrimitiveEngine(engine).allocate(pid, owner,  dLiquidity, true, data);
+    function allocateFromMargin(
+        bytes32 pid,
+        address owner,
+        uint256 dLiquidity,
+        bytes calldata data
+    ) public {
+        IPrimitiveEngine(engine).allocate(pid, owner, dLiquidity, true, data);
     }
 
-    function allocateFromExternal(bytes32 pid, address owner, uint dLiquidity, bytes calldata data) public  {
+    function allocateFromExternal(
+        bytes32 pid,
+        address owner,
+        uint256 dLiquidity,
+        bytes calldata data
+    ) public {
         CALLER = msg.sender;
-        IPrimitiveEngine(engine).allocate(pid, owner,  dLiquidity, false, data);
+        IPrimitiveEngine(engine).allocate(pid, owner, dLiquidity, false, data);
     }
 
-    function allocateCallback(uint dRisky, uint dStable, bytes calldata data) public {
+    function allocateCallback(
+        uint256 dRisky,
+        uint256 dStable,
+        bytes calldata data
+    ) public {
         IERC20(risky).transferFrom(CALLER, msg.sender, dRisky);
         IERC20(stable).transferFrom(CALLER, msg.sender, dStable);
     }
 
-    function getPosition(bytes32 pid) public view returns(bytes32 posid) {
+    function getPosition(bytes32 pid) public view returns (bytes32 posid) {
         posid = keccak256(abi.encodePacked(address(this), pid));
     }
 
     function name() public pure returns (string memory) {
-      return "EngineAllocate";
+        return "EngineAllocate";
     }
 }
-
